+++ ext/sigar/src/os/win32/wmi.cpp
@@ -199,7 +199,7 @@ extern "C" int sigar_proc_args_wmi_get(sigar_t *sigar, sigar_pid_t pid,
         return wmi->GetLastError();
     }
 
-    if (FAILED(wmi->GetProcCommandLine(pid, buf))) {
+    if (FAILED(wmi->GetProcCommandLine((DWORD)pid, buf))) {
         status = wmi->GetLastError();
     }
     else {
@@ -225,7 +225,7 @@ extern "C" int sigar_proc_exe_wmi_get(sigar_t *sigar, sigar_pid_t pid,
 
     procexe->name[0] = '\0';
 
-    if (FAILED(wmi->GetProcExecutablePath(pid, buf))) {
+    if (FAILED(wmi->GetProcExecutablePath((DWORD)pid, buf))) {
         status = wmi->GetLastError();
     }
     else {
